Upgrade Go to 1.23 (#6647)
* Update some for loops in response to the following change introduced
  in Go 1.22: each iteration of the loop creates new variables, to avoid
  accidental sharing bugs. As a result of this change, a few existing
  loop variables had become heap-allocated, so we updated the loop to
  avoid the extra allocation. Other loops that were affected by the
  change (compiling differently), but for which the loop variable was
  stack-allocated, were reviewed individually and when it made sense,
  the loop was refactored to avoid the extra variable copy. The `go test
  -gcflags=-d=loopvar=2` command was used to find affected loops.
* The antrea/codegen image was updated to use Go 1.23. No changes in
  generated files.
* golangci-lint was updated to the latest available version (v1.60.3)

    - govet reported some invalid usages of printf, which was addressed;
      this lead to some refactoring of the `antctl check` code
    - a new gosec rule (G115), which looks for potential integer
      overflows when converting between integer types, caused a lot of
      errors to be reported in the codebase. There were too many errors
      to be addressed as part of this PR, so we only handled a few of
      them for which the fix was straightforward. When the rule
      implementation is improved (so that it reports fewer false
      positives), we should review all errors and address them as needed.

Fixes #6644

+++ pkg/ovs/ovsctl/ovsctl_test.go
@@ -402,7 +402,7 @@ func TestOfCtl(t *testing.T) {
 			bridge:         "br-int",
 			ovsOfctlRunner: mockOVSOfctlRunner,
 		}
-		groupID := 3
+		groupID := uint32(3)
 		DumpGroup := func(id string) ([]byte, error) {
 			var x = []byte{}
 			for i := 0; i < len(testDumpGroups); i++ {
@@ -415,7 +415,7 @@ func TestOfCtl(t *testing.T) {
 			return x, nil
 		}
 		mockOVSOfctlRunner.EXPECT().RunOfctlCmd("dump-groups", strconv.FormatUint(uint64(groupID), 10)).Return(DumpGroup(fmt.Sprintf("%d", groupID)))
-		out, err := client.DumpGroup(uint32(groupID))
+		out, err := client.DumpGroup(groupID)
 		require.NoError(err)
 		expectedGroup := "group_id=3,type=select,bucket=bucket_id:1,output:1,bucket=bucket_id:2,output:2,bucket=bucket_id:3,output:3,bucket=bucket_id:4,output:4"
 		assert.Equal(expectedGroup, out)
